Add correction to finance_calculator.py

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,6 @@
 monthly_income = input("Enter your monthly income: ")
 monthly_expenses = input("Enter your total monthly expenses: ")
-monthly_savings =int(monthly_income) - int(monthly_expenses)
+monthly_savings = float(monthly_income) - float(monthly_expenses)
 
 rate = 0.05
 savings_projection = monthly_savings * 12 + (monthly_savings * 12 * rate)
